FIX: atualização das requisições da api
